Move and refactor problem and render resources
