+++ lib/jahuty/resource/render.rb
@@ -1,0 +1,19 @@
+module Jahuty
+  module Resource
+    class Render
+      attr_accessor :content
+
+      def initialize(content:)
+        @content = content
+      end
+
+      def self.from(payload)
+        Render.new(**payload)
+      end
+
+      def to_s
+        @content
+      end
+    end
+  end
+end
