+++ spec/jahuty/resource/render_spec.rb
@@ -1,19 +1,19 @@
 module Jahuty
-  module Data
+  module Resource
     RSpec.describe Render do
       describe ".from" do
-        let(:data) { { content: "foo" } }
+        let(:payload) { { content: "foo" } }
 
         it "raises error if :content key does not exist" do
-          data.delete(:content)
+          payload.delete(:content)
 
-          expect{ Render.from(data) }.to raise_error(ArgumentError)
+          expect{ Render.from(payload) }.to raise_error(ArgumentError)
         end
       end
 
       describe "#to_s" do
         it "returns content" do
-          expect(Render.new("foo").to_s).to eq("foo")
+          expect(Render.new(content: "foo").to_s).to eq("foo")
         end
       end
     end
